+++ R/textstat.R
@@ -125,15 +125,15 @@ setMethod("[[", "textstat", function(x, i){
 #' @exportMethod [
 #' @importFrom data.table key
 setMethod("[", "textstat", function(x, i, j){
-  if (nrow(x@stat) == 0) warning("Indexing is pointless because data.table is empty")
-  
+  if (nrow(x@stat) == 0L) warning("Indexing is not possible because data.table is empty.")
+
   # Note that i cannot be a call/expression (such as word %in% c("price", "revenue"))
   # in the context of a S4 method
-  
+
   if (is.character(i) && is.null(key(x@stat))){
     if (x@p_attribute %in% colnames(x@stat)) setkeyv(x@stat, cols = x@p_attribute)
   }
-  
+
   if (missing(j)){
     x@stat <- x@stat[eval(i, envir = x@stat)]
     return(x)
@@ -141,6 +141,7 @@ setMethod("[", "textstat", function(x, i, j){
     return( x@stat[i,j, with = FALSE] )
   }
 })
+
 
 setAs(from = "textstat", to = "htmlwidget", def = function(from){
   DT::datatable(from@stat, options = list(pageLength = getOption("polmineR.pagelength"), lengthChange = FALSE))